+++ src/callback.rs
@@ -1,6 +1,5 @@
 use std::{
     fmt::Debug,
-    mem,
     sync::{Arc, Mutex},
 };
 
@@ -42,7 +41,7 @@ impl<T> Callback<T> {
         &self,
         f: Box<dyn FnMut(T) + Send>,
     ) -> Result<OptionBox<dyn FnMut(T) + Send>> {
-        Ok(mem::replace(&mut *self.0.lock()?, Some(f)))
+        Ok(self.0.lock()?.replace(f))
     }
 }